Move `createHangingInputAbortSignal` into its own module
Turbopack doesn't treeshake properly, so this ended up triggering errors
for `process.nextTick` usage in edge bundles.
